Rework del_archives into archive manager
