+++ components/package-template/src/sbin/admin-tools/archive-manager.sh
